+++ ReleaseNotes.md
@@ -21,7 +21,8 @@
 
 ### Continuous Integration
 - cta/CTA#278 - Updating system tests for new EOS evict command
-- cta/CTA#380 - Fix systemtests performance 
+- cta/CTA#380 - Fix systemtests performance
+- cta/CTA#421 - Updated FST GC for new 'eos evict' command
 - cta/CTA#504 - Update CI scripts to use CERN's k8s infrastructure
 - cta/CTA#555 - Upgrade eos-5 to 4.8.105/5.2.4
 - cta/CTA#560 - Fix cppcheck updated errors